+++ NEWS.md
@@ -1,10 +1,16 @@
-# retail 0.0.7
+# trundler 0.0.7
 
 * Add filter by product name in retailer_products().
 * Add filter by product brand in retailer_products().
 
-# retail 0.0.6
+# trundler 0.0.6
 
 * Return 'model' field in retailer_products() and product().
 * Stop paginate() when received fewer rows than requested.
 * Only touch User-Agent in one place.
+
+# trundler 0.1.11
+
+* Deal with possible missing categories in categories().
+* HEAD functionality in products() and product_prices() for API endpoints to only return number of results.
+* Submission to CRAN.